requirements.txt
Flask==0.10.1
Flask-Admin==1.2.0
Flask-Login==0.2.11
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-WTF==0.12
Jinja2==2.8
MarkupSafe==0.23
MultipartPostHandler==0.1.0
Pillow==2.9.0
PyYAML==3.11
SQLAlchemy==1.0.8
WTForms==2.0.2
Werkzeug==0.10.4
dominate==2.1.12
itsdangerous==0.24
wsgiref==0.1.2
